requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
click==8.1.7
fastapi==0.110.0
fastecdsa==2.3.2
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
idna==3.6
Jinja2==3.1.3
leveldb==0.201
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pydantic==2.6.3
pydantic_core==2.16.3
redis==5.0.2
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
torch==2.2.1
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
websockets==12.0
asttokens==2.4.1
base58==0.2.5
certifi==2023.11.17
charset-normalizer==2.0.12
colorama==0.4.6
executing==2.0.1
icecream==2.1.3
pygments==2.17.2
six==1.16.0
pickleDB~=0.9.2
pymongo==4.6.2
python-dotenv==1.0.1